Add files via upload
Provided is a collection of data from a company wishing to process its customer 
records for business purposes (acw_user_data.csv). The existing systems in-place at the 
company only export to a CSV file, and this is not in an appropriate format for analysis. The task is to prepare this data for further analyses by colleagues 
within the company, including representation changes, filtering, and deriving some new 
attributes / metrics for them. 
These data include attributes such as first name, second name, credit card number, marital
status, and even contains data on the customer’s car. 

The number of records provided is significant, and therefore it is expected that solutions are 
robust to varying types of data, and varying values, offering a programmatic solution.